reorganize tab panel content
