+++ frontends/main/src/app-pages/DashboardPage/DashboardPage.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import React, { Suspense } from "react"
+import React from "react"
 import {
   RiAccountCircleFill,
   RiDashboardLine,
@@ -19,31 +19,20 @@ import {
   TabPanel,
   TabList,
   Typography,
-  TypographyProps,
   styled,
 } from "ol-components"
-import { ButtonLink } from "@mitodl/smoot-design"
 import Link from "next/link"
 import { useUserMe } from "api/hooks/user"
 import { useParams } from "next/navigation"
 import UserListListingComponent from "@/page-components/UserListListing/UserListListing"
 import backgroundImage from "@/public/images/backgrounds/user_menu_background.svg"
-import { ProfileEditForm } from "./ProfileEditForm"
+import { ProfileContent } from "./ProfileContent"
 import { useProfileMeQuery } from "api/hooks/profile"
-import {
-  TopPicksCarouselConfig,
-  TopicCarouselConfig,
-  NEW_LEARNING_RESOURCES_CAROUSEL,
-  POPULAR_LEARNING_RESOURCES_CAROUSEL,
-  CERTIFICATE_COURSES_CAROUSEL,
-  FREE_COURSES_CAROUSEL,
-} from "../../common/carousels"
-import ResourceCarousel from "@/page-components/ResourceCarousel/ResourceCarousel"
-import UserListDetailsTab from "./UserListDetailsTab"
-import { SettingsPage } from "./SettingsPage"
+import { UserListDetailsContent } from "./UserListDetailsContent"
+import { SettingsContent } from "./SettingsContent"
 import { DASHBOARD_HOME, MY_LISTS, PROFILE, SETTINGS } from "@/common/urls"
 import dynamic from "next/dynamic"
-import { ResourceTypeEnum } from "api"
+import { HomeContent, TitleText } from "./HomeContent"
 
 const LearningResourceDrawer = dynamic(
   () =>
@@ -219,55 +208,6 @@ const TabPanelStyled = styled(TabPanel)({
   width: "100%",
 })
 
-const TitleText = styled(Typography)<Pick<TypographyProps, "component">>(
-  ({ theme }) => ({
-    color: theme.custom.colors.black,
-    paddingBottom: "16px",
-    ...theme.typography.h3,
-    [theme.breakpoints.down("md")]: {
-      ...theme.typography.h5,
-    },
-  }),
-)
-
-const SubTitleText = styled(Typography)(({ theme }) => ({
-  color: theme.custom.colors.darkGray2,
-  ...theme.typography.body1,
-  [theme.breakpoints.down("md")]: {
-    ...theme.typography.subtitle3,
-  },
-}))
-
-const HomeHeader = styled.div(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  alignSelf: "stretch",
-  [theme.breakpoints.down("md")]: {
-    paddingBottom: "8px",
-  },
-}))
-
-const HomeHeaderLeft = styled.div({
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "flex-start",
-  flex: "1 0 0",
-})
-
-const HomeHeaderRight = styled.div(({ theme }) => ({
-  display: "flex",
-  [theme.breakpoints.down("md")]: {
-    display: "none",
-  },
-}))
-
-const StyledResourceCarousel = styled(ResourceCarousel)(({ theme }) => ({
-  padding: "40px 0",
-  [theme.breakpoints.down("md")]: {
-    padding: "16px 0",
-  },
-}))
-
 const TabKeys = {
   [DASHBOARD_HOME]: "home",
   [MY_LISTS]: "my-lists",
@@ -361,9 +301,6 @@ const DashboardPage: React.FC = () => {
       ? appRouterPath
       : DASHBOARD_HOME
 
-  const topics = profile?.preference_search_filters.topic
-  const certification = profile?.preference_search_filters.certification
-
   const desktopMenu = (
     <ProfileSidebar>
       <Card.Content>
@@ -422,76 +359,12 @@ const DashboardPage: React.FC = () => {
               </DashboardGridItem>
               {showUserListDetail ? (
                 <DashboardGridItem>
-                  <UserListDetailsTab userListId={id} />
+                  <UserListDetailsContent userListId={id} />
                 </DashboardGridItem>
               ) : (
                 <DashboardGridItem>
                   <TabPanelStyled value={DASHBOARD_HOME}>
-                    <HomeHeader>
-                      <HomeHeaderLeft>
-                        <TitleText component="h1">
-                          Your MIT Learning Journey
-                        </TitleText>
-                        <SubTitleText>
-                          A customized course list based on your preferences.
-                        </SubTitleText>
-                      </HomeHeaderLeft>
-                      <HomeHeaderRight>
-                        <ButtonLink variant="tertiary" href={PROFILE}>
-                          Edit Profile
-                        </ButtonLink>
-                      </HomeHeaderRight>
-                    </HomeHeader>
-                    <Suspense>
-                      <StyledResourceCarousel
-                        titleComponent="h2"
-                        title="Top picks for you"
-                        isLoading={isLoadingProfile}
-                        config={TopPicksCarouselConfig(profile)}
-                        data-testid="top-picks-carousel"
-                      />
-                      {topics?.map((topic, index) => (
-                        <StyledResourceCarousel
-                          key={index}
-                          titleComponent="h2"
-                          title={`Popular courses in ${topic}`}
-                          isLoading={isLoadingProfile}
-                          config={TopicCarouselConfig(topic, [
-                            ResourceTypeEnum.Course,
-                          ])}
-                          data-testid={`topic-carousel-${topic}`}
-                        />
-                      ))}
-                      {certification === true ? (
-                        <StyledResourceCarousel
-                          titleComponent="h2"
-                          title="Courses with Certificates"
-                          isLoading={isLoadingProfile}
-                          config={CERTIFICATE_COURSES_CAROUSEL}
-                          data-testid="certification-carousel"
-                        />
-                      ) : (
-                        <StyledResourceCarousel
-                          titleComponent="h2"
-                          title="Free courses"
-                          isLoading={isLoadingProfile}
-                          config={FREE_COURSES_CAROUSEL}
-                          data-testid="free-carousel"
-                        />
-                      )}
-                      <StyledResourceCarousel
-                        titleComponent="h2"
-                        title="New"
-                        config={NEW_LEARNING_RESOURCES_CAROUSEL}
-                        data-testid="new-learning-resources-carousel"
-                      />
-                      <StyledResourceCarousel
-                        titleComponent="h2"
-                        title="Popular"
-                        config={POPULAR_LEARNING_RESOURCES_CAROUSEL}
-                        data-testid="popular-learning-resources-carousel"
-                      />
-                    </Suspense>
+                    <HomeContent />
                   </TabPanelStyled>
                   <TabPanelStyled value={MY_LISTS}>
                     <UserListListingComponent title="My Lists" />
@@ -502,7 +375,7 @@ const DashboardPage: React.FC = () => {
                       <Skeleton variant="text" width={128} height={32} />
                     ) : (
                       <div id="user-profile-edit">
-                        <ProfileEditForm profile={profile} />
+                        <ProfileContent profile={profile} />
                       </div>
                     )}
                   </TabPanelStyled>
@@ -512,7 +385,7 @@ const DashboardPage: React.FC = () => {
                       <Skeleton variant="text" width={128} height={32} />
                     ) : (
                       <div id="user-settings">
-                        <SettingsPage />
+                        <SettingsContent />
                       </div>
                     )}
                   </TabPanelStyled>
